stream/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==40.0.2
discord.py==2.3.2
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pycparser==2.21
python-dotenv==1.0.0
Werkzeug==2.3.7
yarl==1.9.2
